// src/components/interfaces.ts
export interface DisplayProps extends HTMLAttributes<HTMLDivElement> {
    as?: ElementType;
    inline?: boolean;
    pointerEvents?: "none" | "all" | "auto";
    position?: CSSProperties["position"];
    overflow?: CSSProperties["overflow"];
    overflowX?: CSSProperties["overflowX"];
    overflowY?: CSSProperties["overflowY"];
    transition?:
        | "micro-short"
        | "micro-medium"
        | "micro-long"
        | "macro-short"
        | "macro-medium"
        | "macro-long";
    opacity?: 0 | 10 | 20 | 30 | 40 | 50 | 60 | 70 | 80 | 90 | 100;
    zIndex?: -1 | 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10;
}